requirements.txt
torch
torchvision
fastapi
pillow
uvicorn
python-multipart